feat: adding command to get the base data from graphql request

+++ core/models.py
@@ -67,6 +67,6 @@ class Climates(models.Model):
 
 class Planets(models.Model):
     name = models.CharField(max_length=255, unique=True)
-    population = models.IntegerField(blank=True, null=True)
+    population = models.BigIntegerField(blank=True, null=True)
     terrains = models.ManyToManyField(Terrain, blank=True, related_name="planets")
     climates = models.ManyToManyField(Climates, blank=True, related_name="planets")
